updated gameArea, checkLocation and CheckXYcoords to get info from the db.
the app now gets the character coords from the db, and checks those coords against the click. I reworked handleCheckXYcoords in checklocation.js, made it a little easier to read, and used obj name proper { charCoords: charCoords } => { charcoords }, not sure why i didn't do that intitially, but it's done now lol. I also had to change one name in checkXYcoords.js since one of the properties in handleCheckXYcoords in checklocation.js was named differently that what was orginally passed to it.