changing how we check for win.
initially I was checking for win in a useEffect. But due to updating the database for updateCharFound(), sometimes the update would take longer than checkDBForWin(). So it was inconsistent. Now I made it so updateCharacterFound is now async, and it awaits updateDB, then checkDBForWin. This way i think at least it's 100% going to go in the correct order.

so this means we will be checking every time a char is found/updated, instead of on gamedata.characters being updated via useEffect.