fixed a bug on app.js and worked on adding the tempUser to highscores (which works)
so I had been creating the new user gameData object in a useEffect, so it was only being done on the first initial render, not when the game was reset. I moved the createStaticUserInfo to startgame, now it should reset as intended.

in firestore.js I am able to make a copy of the current user to highscores, but it just sends the data, it doesn't add a name, sort by time or do any of the cool stuff I want it to yet.