working on highscores, setTiming final time, and overall end screen (winscreen.js, app.js and firestore.js)
In firestore.js, pushcopyCorrectly was pushing an object holding an object. So it was user: {time: x} which meant I could not orderBy from the firestore query. Time was a nested field. So i fixed that by doing addDoc {...user} instead of addDoc { user }. This seems to have done the trick, and I can now orderBy the time field in getHighscores. Which is awesome

I also have a minor problem that the time displayed is slightly off. GameData time and db times vary by half a second or less. I don't think it's a big deal and i'll chalk it up to the in game timer rounding up and the final display being pinpoint accurate to the milisecond. So now I just need to update the winscreen to display the time from the database.