a lot has changed. Mostly regarding highscores and win screen
allow the user to edit and submit a new name.

displays highscores for the given game!

diplays if you made the highscores table or not!

allows submit until submit/play game? is pressed.

updated some styling on winscreen, just basic stuff. like flexbox , spacing and color.

removed part of the code where we updated highscores prematurely in makeNewHighscores in app.js (it was adding our user data to highscores twice. Once when we beat the game, and once when we submitted/play again (ed) )