changed preview.css
working mostly on the preview image and preview-img-container. I was trying to get the images to look the same without stretching either one, but I settled on going with img-2 being smaller than img1, I can't figure out how to keep the aspect ratios with how different the two images are in size.